requirements.txt
Django==4.2.11
psycopg2-binary==2.9.9
django-widget-tweaks==1.5.0
django-crispy-forms==2.4
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.2
scikit-learn==1.4.1.post1
crispy-bootstrap5==2025.6